// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      user_profiles: {
        Row: {
          company: string | null;
          created_at: string | null;
          email: string | null;
          email_notifications: boolean | null;
          full_name: string | null;
          id: string;
          language: Database["public"]["Enums"]["app_language"] | null;
          push_notifications: boolean | null;
          role: string | null;
          theme: Database["public"]["Enums"]["app_theme"] | null;
          updated_at: string | null;
          weekly_digest: boolean | null;
        };
        Insert: {
          company?: string | null;
          created_at?: string | null;
          email?: string | null;
          email_notifications?: boolean | null;
          full_name?: string | null;
          id: string;
          language?: Database["public"]["Enums"]["app_language"] | null;
          push_notifications?: boolean | null;
          role?: string | null;
          theme?: Database["public"]["Enums"]["app_theme"] | null;
          updated_at?: string | null;
          weekly_digest?: boolean | null;
        };
        Update: {
          company?: string | null;
          created_at?: string | null;
          email?: string | null;
          email_notifications?: boolean | null;
          full_name?: string | null;
          id?: string;
          language?: Database["public"]["Enums"]["app_language"] | null;
          push_notifications?: boolean | null;
          role?: string | null;
          theme?: Database["public"]["Enums"]["app_theme"] | null;
          updated_at?: string | null;
          weekly_digest?: boolean | null;
        };
        Relationships: [];
      };
// ...
    };
// ...
    Enums: {
      app_language: "en" | "es" | "fr" | "de";
      user_role: "user" | "admin";
      app_theme: "light" | "dark" | "system";
      sop_category:
        | "Marketing"
        | "HR"
        | "Operations"
        | "Finance"
        | "Customer Service"
        | "IT"
        | "Sales"
        | "Quality Assurance";
      subscription_plan: "Free" | "Pro" | "Team";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DefaultSchema = Database[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;